Bug #22687023   GENERATED COLUMNS DON'T WORK WITH FOREIGN KEY ACTIONS
Problem:
=======
Foreign key constraint is on a base column of stored column and it is
not affecting the stored column value when foreign key
referential actions are involved.

Fix:
===
Block the DDL which allows to create foreign key constraint on
base column of stored columns when foreign key referential actions
are involved.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
RB: 12335