Bug#23236320: MYSQLTEST DOESN'T NORMALIZE THE CHARACTER_SETS_DIR PATH               IN ITS OUTPUT ON WINDOWS
Issue :
-------
The mysqltest client does not include CHARACTER_SETS_DIR path in
the list of patterns matched and normalized if found in the output.
This may lead to a result mismatch with MTR expecting a normalized
path with only forward slashes but getting a path with backslashes
for the charsets directory on windows.

Fix :
-----
Adding the path of CHARACTER_SETS_DIR to the list of patterns which
are matched in mysqltest output on windows and normalized.

Reviewed by :
Pavan Naik   <pavan.naik@oracle.com>
RB #12614