BUG#23037025 GCOL+PARTITION SIG11 IN ROW_SEL_STORE_MYSQL_REC
Problem :
--------
The root cause for the issue is in fix for bug#21864838/rb#10686.
Partitioning needs the materialized index keys for secondary index
scan. The condition check (m_read_virtual_key) in
row_sel_store_mysql_rec() is not complete and it tries to copy other
non key virtual columns also which may not have been materialized.

Another Bug#22444530 was filed for the same and was fixed by
materializing the non-key columns in row_sel_store_mysql_rec().

This fix resulted in the current bug/crash when SQL returns error
in handler::my_eval_gcolumn_expr() if the base column of a "NOT NULL"
virtual column is NULL.

1. Currently "INSERT IGNORE" allows inserting NULL for base column
where a dependent virtual column is "NOT NULL"

2. thd->count_cuted_fields state controlls whether the virtual column
evaluation would return Error or not. Server sets this state to
CHECK_FIELD_IGNORE for select operations and CHECK_FIELD_ERROR_FOR_NULL
for Insert operation. So, when Insert trigger does a select operation
we get this error while calling handler::my_eval_gcolumn_expr() callback.

3. row_sel_store_mysql_rec() doesn't handle the error and crashes.

Solution :
----------
1. In row_sel_store_mysql_rec, if the virtual column is not materialized
don't attempt to materialize it as it is not required by partitioning.

2. Assert that the column is not part of the index key.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>

RB: 12524