Bug#21098142 7.4 MYSQLD HANGING IN SHUTDOWN
When we do a "shutdown" in mysqld, a deadlock occurs if there is a thread
executing ndb_index_stat at the same time.

The ndb_index_stat thread first takes a local lock and then tries to
acquire a global lock. At the same time, the  shutdown thread takes the
same global lock first and then tries to acquire the same local lock.

Fixed the index stat code to separate the critical sections for the global
lock and the local lock. Now the local lock is released before taking the
global lock, so the deadlock is avoided.