WL#6378: New data dictionary
When replicating from a master with l_c_t_n == 0 to a
slave with l_c_t_n == 1, the schema name is used on the
slave to lookup the default db collation. The schema name
received form the master may contain uppercase letters,
and when using this name, a matching schema object is not
found in the dd.schemata table. Thus, the default server
collation is used instead.

To fix this issue, we instead use the name from 'thd->db()',
which has been lowercased if this is needed.

This patch also changes an enumeration to be strongly typed,
to allow forward declaration and avoiding header file
inclusion in other header files.