Bug#23265447 TEST_EVENT -N CREATEDROPEVENT_NF T1 [DBDICT (LINE: 18834)]
The test 'test_event -n createDropEvent_NF T1' is expected to verify that
dropEvent succeed even if master node crash at a specific time.

This is tested by triggering a crash in Dbdict::execSUB_REMOVE_REF by an
error 6125 insertion.

Unfortunately it seems that the master node do not crash, or not seen
crashing, before the test have ended.

And then occasionally some other test that runs later fails instead.

This patch adds a wait for the master node to stop during the test, this
ensures that following test runs will not crash due to a late crash from
error 6125 insertion.

There can still be a failure in the test itself, this is not examined
further in this bug which focus on removing the effect on other unrelated
test runs.