BUG#22268777 BINLOGLESS SLAVE DOES NOT UPDATE EXEC_MASTER_LOG_POS WITH SLAVE_SKIP_ERRORS
Problem:
In a binlogless slave, the Exec_Master_Log_Pos is not
updated when the error of a statement is ignored by
--slave-skip-errors option.

Analysis:
If the GTID_NEXT was not specified, a binlogless slave
would never update its gtid state when rolling back from
a single statement transaction. The Exec_Master_Log_Pos
would not be updated because, even though the transaction
was finished, its gtid state would show otherwise.

Fix:
Removed the restraint of updating the gtid state when a
transaction is rolled back only if the GTID_NEXT was
specified.