Bug#22576241 SETTING INNODB_MONITOR_ENABLE TO ALL DOES NOT ENABLE ALL MONITORS
Issue:
======
Setting innodb_monitor_enable to ALL does not effectively enable all
monitor counters. The problem is that when innodb_monitor_enable is set to
ALL it enables all monitor counters but does not enable each individual
monitor module.

Yet, in some parts of the code, like in buf_page_monitor, the counters are
only updated if the module is enabled. In particular, this appears to
affect all counters in the buffer_page_io module.

Fix:
====
Now, setting the innodb_monitor_enable to ALL enables the buffer_page_io
counter module, thereby enabling all the counters that belongs to this
module.

RB: 12485
Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>