BUG#23258782 - UPDATE ON P_S TABLES ARE DISALLOWED IN RUNNING GR DUE TO ER3098
[ SERVER SIDE]

The problem with this bug is that though performance schema tables are
not logged in the binary log i.e. they will not be passed to the other
member of the group, but due to the before_dml hook, it will still error
out due to the missing primary key and the non-transactional behavior.

Added a check to disabling any check on the table to see if they are to
be replicated. If they are not to be replicated we ignore such tables.