Bug #23225849: NDB_JOIN_PUSHDOWN* TESTS FAILING ON 32-BIT MACHINES
Problem:

The ndb_join_pushdown* tests began failing on 32-bit machines after
the recent server merge. The range optimization step is skipped
in these tests. In commit 87152b59b563e2749f6ea21a772d9e1d4c6dc258,
the 'range_optimizer_max_mem_size' was lowered to ensure that the
limit was exceeded and the optimization step was skipped on 32-bit
machines as well. However, this change was effectively undone in
commit 4cdb516fe1261b601d48d8d30916299c1e68e37a in which the memory
usage for range optimization was reduced.

Fix:

Further lower the limit such that it is triggered on 32-bit machines
even in the light of the reduction of memory usage for range
optimization.