Bug #23259754 SYS_VARS.INNODB_NUMA_INTERLEAVE_BASIC.TEST NEVER RUNS
Problem:

Binaries are not built with NUMA feature even if it is available and
requested by the user.  This is a regression caused by

commit 73b6f6dfc7056358414cd3dc72cbda6fa523635f
Author: Vasil Dimov <vasil.dimov@oracle.com>
Date:   Tue Jul 21 11:03:25 2015 +0300

Solution:

The WITH_NUMA is a cmake variable and is not available as a macro
to the compiler.  So we should not be using it in #if or #ifdef.
Instead use WITH_NUMA to control the value of HAVE_LIBNUMA, which
can be used for conditional compilation.

Also, an unwanted error log message is removed.

rb#12650 approved by Shaohua.