Bug#22758238 	'WAITING MAX 119 SEC FOR DISTRIBUTING' ON PARALLEL MYSQLD ACTIONS
MySQL 5.7.6 has changed to use 'mysqld --initialize' to create the
datebase schemas. That replaced the previous 'mysqld --bootstrap' which
started a minimal mysqld instance *without a running ndbcluster* and
used a script to create the schemas.

With the introduction of 'mysqld --initialize' the schemas were created
while connected to ndbcluster. Thus the schema operations were also
distributed. However, when initializing a database, schema distribution
is not required as the database schema must be created on all mysqlds
prior to normal startup.

The distribution of schema operations brought us into a MDL-deadlock. (Meta Data Lock)
The root case for this deadlock still remains to be investigated.

As a stopgap for fixing that bug, this patch ensures that a mysqld
started with the --initialize option does not enable the plugin
for the ndbcluster storage engine. Thus initialiuze of database
schemas will not be distributed among the mysqld, which is similar
to the pre 5.7.6 'bootstrap' behavior.