BUG#23194259 ASSERTION FAILURE: TRX0TRX.H:1416:SRV_READ_ONLY_MODE
Problem :
---------
1. Some functions like sha(uuid_short()) is getting treated as const
even if it would evaluate to different value at multiple execution.

2. At prepare stage, in SELECT_LEX::prepare, all partitions are pruned
when sha(uuid_short()) in where condition evaluates to FALSE;

3. index_init in innopart handler is skipped as there is no partition
to scan.

4. At optimize stage, in SELECT_LEX::optimize, sha(uuid_short()) is
evaluated again and it evaluated to TRUE, causing the where condition
to be removed.

5. When pruning is attempted again, in optimize stage, it reset the
part_info BITMAP to scan all partitions as there is no condition.

6. Index scan fails in innopart handler as it is not initialized.

Solution :
----------
We should not reset the part_info BITMAPS during pruning if pruning
is already completed at earlier stage.

Reviewed-by: ADITYA A<aditya.a@oracle.com>
Reviewed-by: Olav Sandstaa <olav.sandstaa@oracle.com>

RB: 12592