Bug #23295655:  MYSQLD failure during SHUTDOWN IF SERVER STARTED   WITH AUDIT_LOG AND USER CONNECTED
Altered the stack overflow check in the audit API to use the
current thread's data instead of the one the message is for.
This is predominantly the same thread, but it can be different if
a separate thread is taking actions on behalf of the existing threads.

(cherry picked from commit cde9d45a2d10b0f7777792252d5a9760a6b656a7)