Bug#23314315 BINLOG.BINLOG_PURGE_BINARY_LOGS_STALL IS FAILING ON PB2
Problem:

The test case is failing on PB2 daily but not on PB2 per push runs.

Analysis:

The test case assert that some lines are present on server's error log,
but the lines are generated only in debug enabled binaries.

Fix:

Added "--source include/have_debug.inc" to the test case.