bug#23305626  GR CAN NOT CHECK LOG-SLAVE-UPDATES ON START GR COMMAND
We do not have a member for LOG-SLAVE-UPDATES in Trans_context_info
struct, so that GR can not check the global variable
LOG-SLAVE-UPDATES on starting GR command.

To fix the problem, add a member log_slave_updates into
Trans_context_info struct and initialize it with global
variable LOG-SLAVE-UPDATES on validating plugin requirements.