WL#6591 Restore backup on existing GTID-aware server - make GTID_PURGED settable always
The patch implements the WL agenda.

Currently, it is possible to add GTIDs to GTID_PURGED only when
GTID_EXECUTED is empty, i.e., when restoring a backup on an empty
server. The enables to add GTIDs to GTID_PURGED at any time, even when
GTID_EXECUTED is nonempty.

The syntax of the regular assignment expressed by

   SET GTID_PURGED = "<gtid_set>"

is augmented with an increment-like append via '+' token
optionally specified as the first non-white space char in the beginning
of the gtid set string:

   SET GTID_PURGED = "+<gtid_set>"

The right hand side of the regular assigment must be a superset of the
GTID_PURGED's current value and their set-difference must be disjoint
with @@GLOBAL.GTID_EXECUTED.
The right hand side of the "append" assignment set must be disjoint
with @@GLOBAL.GTID_EXECUTED set.  In neither case it can overlap with
the current owned GTID set.

client/mysqldump is made to generate version-adoptable SET GTID_PURGED
with the new append syntax for the new (this patch and later) server
and the regular assignment syntax for the old (prior the patch)
servers.