Bug #23259754 SYS_VARS.INNODB_NUMA_INTERLEAVE_BASIC.TEST NEVER RUNS
Problem:

Binaries are not built with NUMA feature even if it is available and
requested by the user.  This is a regression caused by (rb#11231):

commit bf050ecc5c80f72ad068b0228324bebed9064def
Author: Annamalai Gurusami <annamalai.gurusami@oracle.com>
Date:   Mon Dec 7 12:40:49 2015 +0530

This patch attempted to use WITH_NUMA for conditional compilation,
which is wrong.

Solution:

The WITH_NUMA is a cmake variable and is not available as a macro
to the compiler.  So we should not be using it in #if or #ifdef.
Instead use WITH_NUMA to control the value of HAVE_LIBNUMA, which
can be used for conditional compilation.

rb#12650 approved by Shaohua and Tor.