Bug #22679209: FULL-TEXT QUERIES WITH ADDITIONAL SECONDARY INDEX 	       GIVES NULL OR ZERO ROWS
Problem:
When a secondary is chosen for a full-text query, optimizer
does not mark the columns needed to evaluate MATCH function.

Analysis:
In Item_func_match::fix_fields(), read_set is not updated
for the columns needed to evaluate. This works fine, if
optimizer chooses fts index to evaluate the query. If
a secondary index is chosen, innodb will not be able to
fetch the fts_doc_id or the columns required to evaluate
MATCH function. This results in inconsistent results.

Solution:
If fts_doc_id column is not present, update the read_set
for the table with columns needed to evaluate MATCH function.