BUG#23005009: NO WHITESPACE BETWEEN TIMESTAMP AND THREADID               COLUMNS
ANALYSIS:
=========
With the general_log enabled, execution of a query that
has 5 or more digits in its thread_id leads to the
timestamp and thread_id fields being glued together in the
general log.

FIX:
====
Added a tabspace between the timestamp and thread_id
columns.

NOTE:
=====
This is a backport of the patch for Bug#22508563 from trunk
to 5.7.