Bug #22965271 NEEDS REBUILD FOR COLUMN LENGTH CHANGE THAT IS PART OF VIRTUAL INDEX
PROBLEM

While doing an inplace ALTER TABLE operartion, we rebuild the template
only if there is table rebuild or add/drop virtual column. In ALTER
cases where there is a length extension of a column and such a column
also is present in an index with virtual column leading to offset shift
of virtual column value , there is no rebuilt of template hence leading
to assertion failure in ASAN builds .

FIX

Is to add a check for cases where the alter leads to column length
change and call for a rebuilt of template in such cases too to rectify
the incorrect calculation of offset for virtual columns and the assert
failure.

Reviewed-by: Jimmy Yang <Jimmy.Yang@oracle.com>
Reviewed-by: Thirunarayanan Balathandayuthapani
	     <thirunarayanan.balathandayuth@oracle.com>
RB: 12613