Bug#23241575 R-TREE PAGE LOCK PREVENTING THE UNDO TO MERGE THE TREE PAGES
We place page lock while doing any search on rtree, to prevent the search
path gets destroyed by other page shrinks. However, in such, we did not
check the trx of the lock and the one who attempts to do the shrink. And in
the undo case, they could be the same, thus unnecessary preventing shrinking
of page.

Reviewed-by: Jimmy Yang<jimmy.yang@oracle.com>

RB: 12637