WL#8697: Support for INVISIBLE indexes
Follow-up fix.

Made sure that TABLE_SHARE::keys_in_use is correctly initialized even in
presence of hidden indexes (such indexes can be added by SE and are not
currently visible to SQL-layer/not present in TABLE_SHARE::key_info).

Test coverage for this change will be available as part of WL7141.