Bug #18401623 AUTOTEST FAIL: TESTOIBASIC -CASE ABCDEFZ
Also fix Bug#18408738 	AUTOTEST FAIL: TESTOIBASIC -CASE HZ

The ATR test testOIBasic.test pick randomn character sets
for the test tables being created. Internally it keeps a
copy of the rows being inserted into the test tables which
is finally compared against the result produced by Mysql Cluster.

For implementing the character compare method, the strnxfrm()
utility was used to produced a transformed (char set independent)
representation of the strings to be compared. Although this
works for some (most of) character sets, the correct way to
compare character strings are the strnncollsp() function.

Usage if this compare function was implemented as an option
controlled with the '-collsp' argument, which by default was 'off'.

This patch removes this argument and change to always using
strnncollsp().

Patch could be verified by using the big5_chinese_ci collation which
fails when compared as xfrm'ed strings.