BUG#23021140 MTR: RECORD RPL_TEST FOR REPLICATION TEST SUITE
Description
-----------
When MTR is executed with the --skip-rpl option, it does not skip all
replication tests.

MTR currently marks a test as an 'rpl_test' only if the test script or any
other script sourced through it sources include/master-slave.inc. This causes
some tests in the replication suite which don't source include/master-slave.inc
to not be marked with the 'rpl_test' flag.

Fix
---
All tests in the replication suite are marked with the 'rpl_test' flag. This
enables MTR's --skip-rpl option to actually skip replication tests.

This patch is contributed by Daniel Black.

Reviewed by:
Horst Hunger <horst.hunger@oracle.com>
RB #12746