BUG#23519893 CHANNEL SERVICE INTERFACE HAS NO FUNCTION TO GET RETRIEVED              GTID SET
Problem:

Some plug-ins (like Group Replication) need to get information about the
retrieved GTID set for a given replication channel.

The server already has a channel service interface, but it does not have a
function to get the retrieved GTID set.

Without such function, the only way of getting this information is querying
the replication performance schema table about the connection status, but
this table might be unavailable if the server be compiled without performance
schema.

Fix:

Create a new function in the channel service interface to get a channel
retrieved_gtid_set.