BUG#23035906 : MTR: MYSQLTEST PARSES QUOTES IN SQL COMMENTS LITERALLY LEADING TO EOL JUNK ERROR
Description :
=============
The mysqltest parser does not treat quotes used inside SQL comments
as part of a comment. It instead matches quotes(['"`]) literally as
if they are a part of the query and throws errors if they are
imbalanced.

For example, if an SQL comment like / Comment with one quote ' /; is
given to a server, it is valid. But mysqltest searches till the next
matching quote is obtained or else throws end of line junk error.

Fix :
=====
Avoid parsing of quotes if they are present inside SQL comments.

Reviewed-by: Sayantan Dutta <sayantan.dutta@oracle.com>
Reviewed-by: Srikanth B R <srikanth.b.r@oracle.com>
RB: 12374