BUG #22496492 - MYSQL-TEST-RUN CONTAINS OBSOLETE OPTIONS FOR SOLARIS DBX
Description:
The test driver mysql-test-run.pl contains some options that refer to Solaris
dbx. These were needed in MySQL 5.6 and earlier, when it was possible to
compile the programs with the native Solaris compiler, whose output is only
compatible with dbx, not gdb.

Starting with MySQL 5.7, the native Solaris compiler can no longer be used.
Because dbx is incompatible with the output from non-native compilers (such
as GCC and Clang), the dbx-related options are useless.
Description:
The test driver mysql-test-run.pl contains some options that refer to Solaris
dbx. These were needed in MySQL 5.6 and earlier, when it was possible to
compile the programs with the native Solaris compiler, whose output is only
compatible with dbx, not gdb.

Starting with MySQL 5.7, the native Solaris compiler can no longer be used.
Because dbx is incompatible with the output from non-native compilers (such
as GCC and Clang), the dbx-related options are useless.

Solution: Remove the instances of dbx used by MTR.

Reviewed by: Bjorn Munch(bjorn.munch@oracle.com)
RB: 11468