Bug#22275357: ORDER BY DOES NOT WORK CORRECTLY WITH GROUPED AVG()               VALUES EXTRACTED FROM JSON
Queries that are grouped on a column of a BLOB-based type, and that
are ordered on the result of the AVG aggregate function, return
results in the wrong order if InnoDB temporary tables are used.
(InnoDB temporary tables are used by default after WL#6737.)

The bug also affects the aggregate functions that calculate variance,
such as VAR_POP and STDDEV_POP.

The problem is that filesort fails to mark the AVG field in the
table's read_set. This has the consequence that the rows are sorted on
a different column than the one that holds the result from the AVG
aggregate function.

The fix is to make Item_sum_num_field (which is the parent class of
Item_avg_field and Item_variance_field) override mark_field_in_map()
and make it add the field to the read_set. This is a partial backport
of WL#8699, which included a fix for this problem in mysql-trunk.