Bug#23014003 MAIN.WL6443_DEPRECATION CANNOT RUN WITH ./MTR --REPEAT=2(OR MORE)
The test 'main.wl6443_deprecation' checks the number of occurrences of a
deprecation warning in the error log but does not truncate the log later.
Hence, repititions of the test pick up old deprecation warnings causing
the test to fail. Clean up is performed in the test by truncating the
error log.

Reviewed-by: Horst Hunger <horst.hunger@oracle.com>
RB: 12745