Bug#22992666: NULL VALUE IN JSON COLUMN CORRUPTS GROUP_CONCAT               WITH ORDER BY CLAUSE
When GROUP_CONCAT with an ORDER BY clause is used on a JSON column,
and one of the rows has a NULL value in the JSON column, the query
could return wrong results.

Analysis:

GROUP_CONCAT(j ORDER BY j) collects all non-NULL values of the column
"j" in a temporary table and sorts them. Then, to build the result
string, it takes the Field object that represents the "j" column, and
for each row changes Field::ptr to point into the row in the sorted
temporary table instead of the base table, so that it can call
Field::val_str() on each of the sorted values.

If the last row read from the base table had NULL in the "j" column,
the Field's m_null_ptr member will still say that there is a NULL
value in the Field, even after Field::ptr has been changed to point to
a non-NULL value in the temporary table.

Field_json::val_str() calls Field::is_null() to check for NULL, and
returns an empty string if it thinks the value is NULL. Since
is_null() incorrectly says that the value is NULL, val_str() returns
an incorrect value, which again makes GROUP_CONCAT return a wrong
result.

Field_json is the only Field that checks is_null() in its val_xxx
functions. All the other Field classes assume that the caller has
already checked that the value cannot be NULL before val_xxx is
called, and they don't check what is_null() returns. This bug
therefore affects the JSON data type only.

Fix:

Make Field_json make the same assumption as the other Field classes in
its val_xxx functions, that the value is non-NULL when val_xxx is
called. And stop asking is_null(), which is unreliable if the Field
object doesn't point into the original table.