Bug #22843444: INSERTING LARGE JSON DATA TAKES AN                INORDINATE AMOUNT OF TIME
ISSUE:
------
While serializing a json array, the number of allocs is
very high. This is because of the number of times
String::append is called, which in turn calls
String::mem_realloc. The allocation for the buffer happens
linearly and takes significant time.

SOLUTION:
---------
The solution is to allocate larger chunks for the buffer
and reduce the total number of allocs.

The fix is a backport of Bug#22239803.