Bug#23544418 RENAME REALFRAGMENT TO PARTITION IN NDB KERNEL CODE.
Post push fix for WL#9019.

This patch is pure refactoring renaming RealFragmentCount to PartitionCount
in kernel code.

One visible change from 7.5.2 is that in output of ndb_desc
RealFragmentCount is now called ParititionCount.

Background

With fully replicated tables the notion of Real Fragments was introduced in
the code, but later these was decided to be called Partitions.

This is a pure refactoring with no user visible changes (except for user
looking in code :)

The nomenclature for partitions and fragments shoud be:

1) partition - the user visible partitions in MySQL.
2) fragment - in a node group it is a logical collection of some rows in a
table. Fragments can not be shared between node groups.
3) fragment replica - each fragment in a node group are stored in one
fragment replica on each node in the node group.
4) primary (fragment) replica - one of nodes in the node group have the
primary replica there all operations needing a lock is directed to.
5) backup (fragment) replica - the fragment replicas that are not the primary
replica.

For traditional tables each fragment corresponds to a partition, but with
fully replicated table there all node groups and nodes have its own copy of
all rows in the table, a partition have one fragment in each node group all
with the same data.

6) copy fragment - a partition consists of possible several copy fragments,
one per node group, all with same data, but each copy fragment have its
unique id.
7) main fragment - for a partition the main fragment is the fragment with the
primary replica used for updates and lock operations.