Fix failing ATR test: 'testSpj -n MixedJoin':
The test utilit function runJoin() had the option to
set the property 'InsertError' which in case specified
were used as ERROR_INSERT value. When runJoin() completed
it *unconditionally* inserted a '0' error in order to clean
any error prev. inserted.

The testcase MixedJoin ran 6 steps of this testcase in parallel
*without* any InsertError property being specified. However,
the last error=0 where nevertheless always set.

The implementation if error insert always send a SYNC_REQ
signal after an error insert in order to 'flush' it
to all Ndb blocks. This signal is routed via the block proxy
which have only 4 slots for outstandig SYNC_REC signal.
If this limit is exceeded it will crash in LocalProxy.hpp line 217.

As an error insert sets the error globally for all clients connected
to the datanode, such parall setting of error inserts have never
been an intended part of what we support. Thus it should be avoided.
(STEPS == 1 whenever an error is inserted)

This patch makes the 0 error insert conditionally, such that
the proxy slots for SYNC_REQ are not exhaused.