Fix spurious failures in ATR test 'testTransactions'
This test runs a mix of 2 different transaction T1/T2. For some
of the testcases, the second transaction is supposed to
be blocked by the first onw , and thus timeout with either
error 266 or 274. For other combinations, no such timeouts are
accepted.

In order to reduce the elapsed time for the test, the TC transaction
timeout was reduced to 100ms (default 3000ms). That had the
negative impact that even transaction combination not expected to
block each other, code spriously timeout.

As the setting of TC txn timeout was a 'global' setting, it
also remained set after this test had completed, and thus
could result in timeouts for later ATR test cases.

This patch:

1) Delay setting of  the 'short timeout' (100ms) to after
   The first (non-blocking( transaction is executed. Thus
   avoiding 'T1' to timeout.

2) Sets the 'short timeout' for T2 only for T1/T2 combinations
   where T2 is expected to fail with a timeout

3) Set transaction timeout back to 'default' as soon as the
   T2 has completed, this avoids that the short timeout
   affects other transactions than the one assumed to timeout.