Fix several failures in 'testSystemRestart -n SR_DD_n*' tests
Tests where inserting into tables or updating rows at the
max speed of the test client machines for the duration of
each test loop.

As the new vigdis servers are fastart than the machines they
replaced, we are now able to fill the TableSpace and/or
redo logs to their max sizes, and thus the test failes.

This patch introduce a load limiting machanism which limit
the client load to 10.000 insert/update/deletes pr sec.

This should also avoid that these tests fails the next time
the ATR test are moved to faster hardware.