Improve comments about assumptions to ROR scans.
Just from a local reading of the code, it was not at all clear that
the ROR intersection plans expect R-B trees to have only a single
element. (It is, of course, clearer if you know exactly what a ROR
intersection plan is, in particular that iff you read a single value
from an InnoDB index, you are guaranteed to get it in the same order
as the main table. But this is a lot of context that may not be known
to the casual reader.)

Improve the comment and add a debug assert to make this clearer.

Change-Id: Ib131c18182e5b3908a65c7a164c626817bd2c485