Bug#22671573: SIG11 IN SEL_ARG::RB_INSERT |SQL/OPT_RANGE.CC
Split out the concept of "a node in the SEL_ARG tree" from the concept
of the tree itself (the latter is now called SEL_ROOT). This untangles
the two concepts (although the split isn't complete; parts of the split
are complicated enough to be future patch); in particular, SEL_ARG
no longer needs to carry around fields which only makes sense at
the root (such as use_count or type), and copy them around when the root moves.

The motivating change, however, is that SEL_ARGs can now point to SEL_ROOTs
instead of another SEL_ARG. Previously, the pointer was implicitly to the
root node of the tree; however, a node that used to be the root can later be
made non-root due to R-B tree rotations, which would then cause issues when
it was attempted traversed. Now, SEL_ROOT is the sole place holding the
SEL_ARG root, which is then easy to keep up to date during rotations.
SEL_ROOT also now holds the "type" field, which uncovered several minor issues
where our refcounting was either too conservative or aggressive.

If we want to, this also opens up for more opportunities to reduce the size
of SEL_ARG down the road; in particular, the "part" and "field" fields
(and possibly "maybe_flag") can be probably moved into SEL_ROOT with a bit of
effort, and the "parent" pointer is now only needed during insertions and
deletions, so can also probably be held implicitly. Optionally, we can
store all SEL_ARGs in a contiguous array in the SEL_ROOT, enabling use of
smaller indexes instead of pointers, and possibly also other traversal
optimizations. Or, if we just want cleanup, we can get rid of the R-B tree
entirely and replace it with a std::map.

Change-Id: Ia152b411d4d4dde98c9a3adb70aad958bccfba81