Bug#23503675: ASSERT IN DATABASE INTIALIZATION DUE TO               INNODB_DEFAULT_ROW_FORMAT=COMPACT
The problem was that the dictionary table for spatial reference
systems (WL#8579) was not constructed with explicit ROW_FORMAT set.
This meant that it was constructed with whatever ROW_FORMAT
was set as default by the server.

This patch fixes the problem by setting ROW_FORMAT=DYNAMIC
explicitly for this dictionary table when it is created -
similarly to what is done for other dictionary tables.

The patch also refactors the code for creating dictionary tables
so that common options (such as ROW_FORMAT) is set in one
place only so that we don't have to remeber doing for any
future dictionary tables added.

The patch also moves the definitions of some large constructors
from header files to cc files.