BUG#19502202 - SERVER SHUTDOWN HANG SEEN IN SOME INNODB                & RPL TESTS
A MTR test binlog.binlog_xa_prepared_disconnect fails sporadically
on PB2. When the server is shutdown during the execution of the tests,
the shutdown of the mysql server doesn't happen within a specified
timeout thereby causing SIGABRT to be sent to the server and the test
aborts.
As part of the shutdown sequence, sockets are shutdown and closed in
one thread and this shall cause the POLLHUP event to be generated for
the sockets that are stuck in poll sleep state on another thread. The
generation of POLLHUP from the kernel is async and not deterministic
and time-bound.  Hence shutdown hangs and could not complete in a
specified timeout period. The patch fixes this by using ppoll in
vio_io_wait and sending pthread_kill to the thread that is stuck
on poll when the socket is undergoing shutdown.
The constructor and destructor functionality to the patch was addded
by Tor Didriksen <tor.didriksen@oracle.com>.