Bug#21221500: ADD SUPPORT FOR GCC __ATOMIC BUILTINS
This patch adds support for the GCC __atomic builtins introduced
in GCC 4.7. These new builtins allows a more efficient
implementation of the MySQL atomics API. Most importantly the
new builtins include functions for atomic load, so that
my_atomic_load can be implemented as a real load instead of a store.

The new __atomic builtins are only used for platforms
where the old GCC __sync builtins are not available.
In particular this includes platforms supported by Ubuntu
(such as powerpc) where the server would otherwise not build.