Bug#23623359: STOP COMPILING EVENT_PARSE_DATA.CC WHEN BUILDING               EMBEDDED SERVER
Move the definition of the Event_parse_data ctor from
the .cc file to the .h file so that the .cc file no longer
needs to be compiled when building embeded server.
(The ctor is needed in sql_yacc.yy).

Also move mysqld_thd_manager.cc to the list of common
source files for normal server and embedded server,
rather than list it separately in each of the servers
unique source file lists.