Bug#23023457	DEBIAN/UBUNTU PACKAGES FAIL TO INSTALL BECAUSE OF EMPTY CONFIG VALUES
Post-push fix:
The test was failing on Windows; Since the client no longer throws an error it would
go on to try to connect to the server, and on Windows the test doesn't set up a
server to listen on the default port.
Since the test only really needs to check that the client prints the warning and does
not throw an error, we change the test to run mysql --help instead of trying to
connect, passing stdout to a temporary file so only the warning is left.