Bug#23607230 WRITE SET HASHES SHOULD HAVE A 64BITS OPTION TO REDUCE WRITE CONFLICTS
As of now, the server only provided a 32 bit option for write set
extraction. This is the algorithm that identifies updates based on
their database, table and rows. In large workloads this can lead to
different updates with the same id.

To solve this, a faster algorithm was introduced that outputs a 64
bits string. This allows a better identification of updates with less
false conflicts.
Small corrections are also made to the input passed to the hash
algorithm.